test(agent-tracing): split the projection tests and tighten two assertions
The Rust layout gate caps a file at 750 lines and
`journal_projection_tests.rs` had reached 755. Move the new #6148 cases into
`journal_projection_tests_part_02_tests.rs`, declared from the parent with
`#[path]` — the same shape `progress_tracing_tests.rs` already uses — leaving
the fixtures and the pre-existing cases where they were.

Two assertions strengthened, both per review:

- The sub-agent `UnknownToolCall` test asserted only that the tool span had
  *some* parent, which would pass even if it attached to an unrelated
  iteration. Assert the exact chain instead: the tool hangs off the child's
  iteration span, which hangs off the sub-agent span.

- The no-op arm test compared span counts and kinds, which does not prove the
  documented "identical tree" — a diagnostic event could still change an
  attribute, name, status or parent link. Compare every stable field via a
  `span_shape` helper that excludes the generated `trace_id`/`span_id` and
  expresses the parent positionally, so the tree shape is pinned without
  depending on id generation.

95 passed / 0 failed across `agent::progress_tracing`; layout gate green.
